backend/requirements.txt
langchain
openai
pinecone
langchain-pinecone
langchain-openai
python-dotenv
pypdf